requirements/common.txt
Django==1.8.2
mysql-python==1.2.5
boto==2.38.0
django-rq==0.7.0    # Async process
djangorestframework==3.1 # REST API
markdown==2.5.2       # Markdown support for the browsable API.
django-rq-dashboard
django-filter==0.10.0  # Filtering support
#https://github.com/fefa4ka/voiceid/archive/master.zip # Before install SOX and GSTREAM
numpy==1.9.1
pydub==0.9.5
django-storages
django-s3-folder-storage
djangorestframework-bulk==0.1.3
django-social-auth
bootstrap-admin1
djoser
diff-match-patch
newrelic
xlsxwriter
django-wpadmin
python-docx
django-db-mailer
django-mailbox
django-redis-cache

#httplib2